+++ src/main/java/com/guxingke/intellij/plugin/Const.java
@@ -2,6 +2,7 @@ package com.guxingke.intellij.plugin;
 
 public interface Const {
 
+  String CLS_JAVA_LANG_STRING = "java.lang.String";
   String CLS_JAVA_UTIL_COLLECTION = "java.util.Collection";
   String CLS_JAVA_UTIL_STREAM_STREAM = "java.util.stream.Stream";
 }