+++ src/main/java/com/guxingke/intellij/plugin/postfix/template/ToMapPostfixTemplate.java
@@ -4,6 +4,7 @@ import com.guxingke.intellij.plugin.Const;
 import com.guxingke.intellij.plugin.util.PsiExpressionUtils;
 import com.intellij.codeInsight.template.TemplateManager;
 import com.intellij.codeInsight.template.impl.TextExpression;
+import com.intellij.codeInsight.template.impl.Variable;
 import com.intellij.codeInsight.template.postfix.templates.PostfixTemplateProvider;
 import com.intellij.codeInsight.template.postfix.templates.PostfixTemplateWithExpressionSelector;
 import com.intellij.codeInsight.template.postfix.util.JavaPostfixTemplatesUtils;
@@ -14,10 +15,10 @@ import com.intellij.psi.PsiExpression;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-public class ToIdentifierMapPostfixTemplate extends PostfixTemplateWithExpressionSelector {
+public class ToMapPostfixTemplate extends PostfixTemplateWithExpressionSelector {
 
-  public ToIdentifierMapPostfixTemplate(@Nullable PostfixTemplateProvider provider) {
-    super("toIdMap", "toIdMap", "convert to id map", JavaPostfixTemplatesUtils.selectorTopmost(cond()), provider);
+  public ToMapPostfixTemplate(@Nullable PostfixTemplateProvider provider) {
+    super("toMap", "toMap", "convert to map", JavaPostfixTemplatesUtils.selectorTopmost(cond()), provider);
   }
 
   private static Condition<PsiElement> cond() {
@@ -55,10 +56,11 @@ public class ToIdentifierMapPostfixTemplate extends PostfixTemplateWithExpressio
       return;
     }
 
+    var common = ".collect(java.util.stream.Collectors.toMap($componentClassName$::$keyF$, $valueF$))$END$";
     var stream = PsiExpressionUtils.isClass(e, "java.util.stream.Stream");
-    var ts = "$expr$.stream().collect(java.util.stream.Collectors.toMap($componentClassName$::getId, it -> it, (l, r) -> l))$END$";
+    var ts = "$expr$.stream()" + common;
     if (stream) {
-      ts = "$expr$.collect(java.util.stream.Collectors.toMap($componentClassName$::getId, it -> it, (l, r) -> l))$END$";
+      ts = "$expr$" + common;
     }
 
     document.deleteString(expression.getTextRange().getStartOffset(), expression.getTextRange().getEndOffset());
@@ -66,6 +68,11 @@ public class ToIdentifierMapPostfixTemplate extends PostfixTemplateWithExpressio
     var tpl = manager.createTemplate(getId(), "", ts);
     tpl.addVariable("expr", new TextExpression(expression.getText()), false);
     tpl.addVariable("componentClassName", new TextExpression(cls.getQualifiedName()), false);
+
+    var fv = new Variable("keyF", "", "", true);
+    tpl.addVariable(fv.getName(), fv.getExpression(), fv.getDefaultValueExpression(), true, false);
+    var vv = new Variable("valueF", "", "", true);
+    tpl.addVariable(vv.getName(), vv.getExpression(), vv.getDefaultValueExpression(), true, false);
     tpl.setToReformat(true);
     manager.startTemplate(editor, tpl);
   }